feat: add exception handler - 예외 처리 로직 추가

+++ src/main/resources/script/threelines_ddl_v1.0.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `MEMBER` (
-                          `MEMBER_NO` INT AUTO_INCREMENT PRIMARY KEY,
+                          `MEMBER_NO` BIGINT AUTO_INCREMENT PRIMARY KEY,
                           `MEMBER_EMAIL` VARCHAR(255) NULL,
                           `MEMBER_NM` VARCHAR(255) NULL,
                           `MEMBER_PW` VARCHAR(255) NULL,
@@ -10,7 +10,7 @@ CREATE TABLE `MEMBER` (
 );
 
 CREATE TABLE `SUBSCRIPTION` (
-                                `SUB_NO` INT AUTO_INCREMENT PRIMARY KEY,
+                                `SUB_NO` BIGINT AUTO_INCREMENT PRIMARY KEY,
                                 `SUB_TITLE` VARCHAR(255) NULL,
                                 `SUB_TYPE` VARCHAR(255) NULL,
                                 `MEMBER_NO` INT NOT NULL,
@@ -21,7 +21,7 @@ CREATE TABLE `SUBSCRIPTION` (
 );
 
 CREATE TABLE `CONTENT` (
-                           `CONTENT_NO` INT AUTO_INCREMENT PRIMARY KEY,
+                           `CONTENT_NO` BIGINT AUTO_INCREMENT PRIMARY KEY,
                            `CONTENT_TITLE` VARCHAR(255) NULL,
                            `CONTENT_ADDR` VARCHAR(255) NULL,
                            `CONTENT_SUMMARY` TEXT NULL,
